pyproject.toml

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eos_gcc"
version = "0.1.0"
description = "GCC End-of-Service engine for Frappe/ERPNext"
authors = [{name="Neotec Integrated Solutions"}]
requires-python = ">=3.10"
dependencies = ["python-dateutil>=2.8.2"]
